ai/EditGuard/requirements.txt
absl-py==1.0.0
addict==2.4.0
aiofiles==23.2.1
alabaster==0.7.12
altair==5.4.1
anyio==4.4.0
apex==0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
basicsr==1.4.2
beautifulsoup4==4.11.1
bleach==5.0.0
blis==0.7.7
brotlipy==0.7.0
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
codecov==2.1.12
colorama==0.4.4
conda==4.12.0
conda-build==3.21.8
conda-package-handling==1.8.1
coverage==6.3.2
cryptography==36.0.2
cuda-python==11.6.1
cudf==22.2.0a0+309.gdad51a548e
cugraph==22.2.0a0+95.g1a25c914
cuml==22.2.0a0+109.gc41beba23
cupy-cuda115==9.6.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
dask==2021.11.2
dask-cuda==22.2.0
dask-cudf==22.2.0a0+309.gdad51a548e
dataclasses==0.8
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.30.2
distributed==2021.11.2
docutils==0.17.1
einops==0.8.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==0.8.3
expecttest==0.1.3
fastapi==0.113.0
fastjsonschema==2.15.3
fastrlock==0.8
ffmpy==0.4.0
filelock==3.6.0
Flask==2.1.1
fonttools==4.32.0
fsspec==2024.9.0
future==0.18.2
glob2==0.7
google-auth==2.34.0
google-auth-oauthlib==1.0.0
gradio==3.50.0
gradio_client==0.6.1
graphsurgeon==0.4.5
grpcio==1.66.1
h11==0.14.0
HeapDict==1.0.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hypothesis==4.50.8
idna==3.3
imageio==2.35.1
imagesize==1.3.0
importlib_metadata==8.4.0
importlib-resources==5.7.0
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyter-tensorboard==0.2.0
jupyterlab==2.3.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupytext==1.13.8
kiwisolver==1.4.2
langcodes==3.3.0
lazy_loader==0.4
libarchive-c==4.0
librosa==0.8.1
llvmlite==0.36.0
lmdb==1.3.0
locket==0.2.1
Markdown==3.3.6
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.0
mdurl==0.1.1
mistune==0.8.4
mock==4.0.3
msgpack==1.0.3
murmurhash==1.0.6
narwhals==1.6.2
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.7
notebook==6.4.1
numba==0.53.1
numpy==1.22.3
nvidia-dali-cuda110==1.12.0
nvidia-pyindex==1.0.9
nvtx==0.2.4
oauthlib==3.2.0
onnx==1.11.0
opencv-python==4.10.0.84
opencv-python-headless==4.2.0.34
orjson==3.10.7
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
pkginfo==1.8.2
platformdirs==4.2.2
pluggy==1.0.0
polygraphy==0.33.0
pooch==1.6.0
preshed==3.0.6
prettytable==3.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycocotools==2.0+nv0.6.0
pycosat==0.6.3
pycparser==2.21
pydantic==1.8.2
pydot==1.4.2
pydub==0.25.1
Pygments==2.11.2
pynvml==11.4.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
python-hostlist==1.21
python-multipart==0.0.9
python-nvd3==0.15.0
python-slugify==6.1.1
pytorch-quantization==2.1.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
revtok==0.0.3
rmm==22.2.0a0+37.gf2b1364
rsa==4.8
ruamel-yaml-conda==0.15.80
sacremoses==0.0.49
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==0.24.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.0
setuptools==59.5.0
shellingham==1.4.0
six==1.16.0
smart-open==5.2.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
Sphinx==4.5.0
sphinx-glpi-theme==0.3
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.3
stack-data==0.2.0
starlette==0.38.4
tabulate==0.8.9
tb-nightly==2.14.0a20230808
tblib==1.7.0
tensorboard==2.8.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorrt==8.2.4.2
terminado==0.13.3
text-unidecode==1.3
thinc==8.0.15
threadpoolctl==3.1.0
tifffile==2023.7.10
tinycss2==1.1.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.12.0a0+bd13bc6
torch-tensorrt==1.1.0a0
torchtext==0.13.0a0
torchvision==0.13.0a0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.44.2
treelite==2.2.1
treelite-runtime==2.2.1
typer==0.4.1
typing_extensions==4.12.2
ucx-py==0.21.0a0+37.gbfa0450
uff==0.6.9
urllib3==1.26.9
uvicorn==0.30.6
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.1.1
wheel==0.37.1
xgboost==1.5.2
yapf==0.40.2
zict==2.1.0
zipp==3.8.0
